move post-processing trex_utils to decoders
